fix dropzone
